Work around a bug in latest pyvis
Latest pyvis introduced the notebook kw parameter to Graph.show(), but
it crashes for us with the default value of notebook=True. To work
around the issue I could either pin the dependency to the previous
version or explicitly set the value to False while enforcing using the
most recent version. THe latter approach has the advantage of being
forward-compatible with future versions of the package (assuming they
don't break API!!)