Add update-excuse bug link when appropriate
This is really hackish, but is a start. Next step would be to query LP
directly to see if the bug is assigned (e.g. is it worked on by anyone),
but I don't have the time to do so just yet.

Note that this is not support for the 'block-proposed' tag, only the
'update-excuse' one.