Make cache_dir mandatory in CachedExcuses and change default
The whole point of the class is to cache information, thus it makes
sense to make the parameter mandatory. This also pushes its use further
"up" to the stack with the goal of exposing it as a command line option.

Secondly (and possibly more controversially), this changes the default
cache location to $XDG_CACHE_HOME/visual-excuses where XDG_CACHE_HOME is
defaulted to "~/.cache". This is inline with the freedesktop standard.