Re-write CachedExcuses.update to avoid TOCTOU
The former version checks whether the etag is up to date with HEAD, then
proceeds with GET if it is not, but uses the etag *from the HEAD fetch*.
This leads to two erroneous possibilities, one of which is trivial and
the other more serious:

Firstly, the resource may appear up to date in the HEAD call, but
immediately after be updated. In this case the script will not refresh
but arguably is still correct because the next fetch will operate
correctly.

Secondly, the resource may appear out of date as the HEAD etag is
different. But before the GET, the resource changes. Now the GET
proceeds but the *wrong etag* is stored.

The solution is simply to perform a GET with the cached etag (if
present), and act according to the resulting status code (304: do
nothing, 200: proceed to download the data, anything else: throw
exception). This commit makes this change, and also avoids storing the
entire download in memory, streaming the decompression onto disk.