Use HEAD+GET
Revert this commit if/when the upstream server gets fixed to honour the
ETag it sends correctly. Currently, the server serves an ETag, but
completely ignores the If-None-Match header, so we have to handle
avoiding the download manually.

This introduces a TOCTOU issue (albeit an unlikely one), but as long as
the server is broken, we've no other option here.